(feat: core) Moved the anthropic conversation into the CLI
