+++ src/bin/passport_for_ai.rs
@@ -1,7 +1,9 @@
 use clap::{Arg, Command};
-use passport_for_ai::local;
+#[cfg(feature = "remote")]
+use passport_for_ai::remote;
 use std::error::Error;
-use std::path::Path;
+#[cfg(feature = "local")]
+use {passport_for_ai::local, std::path::Path};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
@@ -9,147 +11,140 @@ async fn main() -> Result<(), Box<dyn Error>> {
         .version("0.1")
         .about("AI Passport CLI")
         .subcommand(
-            Command::new("create-passport")
-                .about("Creates a unique passport for an ONNX model")
-                .arg(
-                    Arg::new("local")
-                        .long("local")
-                        .help("Indicates that this is for a local model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("remote"),
+            Command::new("local")
+                .about("Operations for local models")
+                .subcommand(
+                    Command::new("create-passport")
+                        .about("Creates a unique passport for a local ONNX model")
+                        .arg(
+                            Arg::new("model_path")
+                                .help("Path to the ONNX model file")
+                                .required(true)
+                                .index(1),
+                        )
+                        .arg(
+                            Arg::new("save_to_path")
+                                .long("save-to")
+                                .help("Optional path to the directory where the passport will be saved")
+                                .value_hint(clap::ValueHint::DirPath),
+                        ),
                 )
-                .arg(
-                    Arg::new("remote")
-                        .long("remote")
-                        .help("Indicates that this is for a remote model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("local"),
+                .subcommand(
+                    Command::new("attribute-content")
+                        .about("Attributes content to the local model")
+                        .arg(
+                            Arg::new("model_path")
+                                .help("Path to the model to attribute the content to")
+                                .required(true)
+                                .index(1),
+                        )
+                        .arg(
+                            Arg::new("content_path")
+                                .help("Path to the content file to attribute")
+                                .required(true)
+                                .index(2),
+                        )
+                        .arg(
+                            Arg::new("save_to_path")
+                                .long("save-to")
+                                .help("Optional path to the directory where the proof will be saved")
+                                .value_hint(clap::ValueHint::DirPath),
+                        ),
                 )
-                .arg(
-                    Arg::new("model_path")
-                        .help("Path to the ONNX model file")
-                        .required(true)
-                        .index(1),
-                )
-                .arg(
-                    Arg::new("save_to_path")
-                        .long("save-to")
-                        .help("Optional path to the directory where the passport will be saved")
-                        .value_hint(clap::ValueHint::DirPath),
+                .subcommand(
+                    Command::new("verify-attribution")
+                        .about("Verifies the cryptographic proof for a local ONNX model")
+                        .arg(
+                            Arg::new("model_passport_path")
+                                .help("Path to the JSON model passport")
+                                .required(true)
+                                .index(1),
+                        )
+                        .arg(
+                            Arg::new("attribution_certificate_path")
+                                .help("Path to the attribution certificate file (attribution_certificate.json)")
+                                .required(true)
+                                .index(2),
+                        ),
                 ),
         )
         .subcommand(
-            Command::new("attribute-content")
-                .about("Attributes content to the model")
-                .arg(
-                    Arg::new("local")
-                        .long("local")
-                        .help("Indicates that this is for a local model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("remote"),
-                )
-                .arg(
-                    Arg::new("remote")
-                        .long("remote")
-                        .help("Indicates that this is for a remote model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("local"),
-                )
-                .arg(
-                    Arg::new("model_path")
-                        .help("Path to the model to attribute the content to")
-                        .required(true)
-                        .index(1),
-                )
-                .arg(
-                    Arg::new("content_path")
-                        .help("Path to the content file to attribute")
-                        .required(true)
-                        .index(2),
-                )
-                .arg(
-                    Arg::new("save_to_path")
-                        .long("save-to")
-                        .help("Optional path to the directory where the proof will be saved")
-                        .value_hint(clap::ValueHint::DirPath),
-                ),
-        )
-        .subcommand(
-            Command::new("verify-attribution")
-                .about("Verifies the cryptographic proof for a provided ONNX model")
-                .arg(
-                    Arg::new("local")
-                        .long("local")
-                        .help("Indicates that this is for a local model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("remote"),
-                )
-                .arg(
-                    Arg::new("remote")
-                        .long("remote")
-                        .help("Indicates that this is for a remote model")
-                        .action(clap::ArgAction::SetTrue)
-                        .conflicts_with("local"),
-                )
-                .arg(
-                    Arg::new("model_passport_path")
-                        .help("Path to the Json model passport")
-                        .required(true)
-                        .index(1),
-                )
-                .arg(
-                    Arg::new("attribution_certificate_path")
-                        .help("Path to the attribution certificate file (attribution_certificate.json)")
-                        .required(true)
-                        .index(2),
+            Command::new("remote")
+                .about("Operations for remote models")
+                .subcommand(
+                    Command::new("anthropic-conversation")
+                        .about("Interact with the Anthropic API to generate a proof of conversation"),
                 ),
         )
         .get_matches();
 
-    if let Some(matches) = matches.subcommand_matches("create-passport") {
-        let model_path = Path::new(matches.get_one::<String>("model_path").unwrap());
-        let save_to_path = matches.get_one::<String>("save_to_path").map(Path::new);
+    #[allow(unused_variables)]
+    // Handle `local` commands
+    if let Some(local_matches) = matches.subcommand_matches("local") {
+        #[cfg(feature = "local")]
+        {
+            if let Some(matches) = local_matches.subcommand_matches("create-passport") {
+                let model_path = Path::new(matches.get_one::<String>("model_path").unwrap());
+                let save_to_path = matches.get_one::<String>("save_to_path").map(Path::new);
 
-        if matches.get_flag("local") {
-            local::create_model_passport(model_path, save_to_path)
-                .await
-                .map_err(|err| format!("Error generating model passport: {err}"))?;
-        } else if matches.get_flag("remote") {
-            eprintln!("Error: Remote models are not implemented yet.");
+                local::create_model_passport(model_path, save_to_path)
+                    .await
+                    .map_err(|err| format!("Error generating model passport: {}", err))?;
+            } else if let Some(matches) = local_matches.subcommand_matches("attribute-content") {
+                let model_path = Path::new(matches.get_one::<String>("model_path").unwrap());
+                let content_path = Path::new(matches.get_one::<String>("content_path").unwrap());
+                let save_to_path = matches.get_one::<String>("save_to_path").map(Path::new);
+
+                local::prove_attribution(model_path, content_path, save_to_path)
+                    .await
+                    .map_err(|err| format!("Error attributing content to the model: {}", err))?;
+            } else if let Some(matches) = local_matches.subcommand_matches("verify-attribution") {
+                let model_passport_path =
+                    Path::new(matches.get_one::<String>("model_passport_path").unwrap());
+                let attribution_certificate_path = Path::new(
+                    matches
+                        .get_one::<String>("attribution_certificate_path")
+                        .unwrap(),
+                );
+
+                local::verify_attribution(model_passport_path, attribution_certificate_path)
+                    .await
+                    .map_err(|err| format!("Error verifying attribution: {}", err))?;
+            } else {
+                eprintln!(
+                    "Error: Invalid local subcommand. Use '--help' to see available commands."
+                );
+                std::process::exit(1);
+            }
+        }
+
+        #[cfg(not(feature = "local"))]
+        {
+            eprintln!("Error: this binary was not compiled with the 'local' feature enabled. Hence, local operations are not supported.");
             std::process::exit(1);
         }
     }
-
-    // Handle `attribute-content` command
-    if let Some(matches) = matches.subcommand_matches("attribute-content") {
-        let model_path = Path::new(matches.get_one::<String>("model_path").unwrap());
-        let content_path = Path::new(matches.get_one::<String>("content_path").unwrap());
-        let save_to_path = matches.get_one::<String>("save_to_path").map(Path::new);
-
-        if matches.get_flag("local") {
-            local::prove_attribution(model_path, content_path, save_to_path)
-                .await
-                .map_err(|err| format!("Error attributing content to the model: {err}"))?;
-        } else if matches.get_flag("remote") {
-            eprintln!("Error: Remote models are not supported yet.");
+    // Handle `remote` commands
+    else if let Some(remote_matches) = matches.subcommand_matches("remote") {
+        #[cfg(feature = "remote")]
+        {
+            if let Some(_) = remote_matches.subcommand_matches("anthropic-conversation") {
+                remote::generate_proof_of_conversation()
+                    .await
+                    .map_err(|err| format!("Error during conversation: {}", err))?;
+            } else {
+                eprintln!("Error: The specified remote feature is not available yet. Currently, only 'anthropic-conversation' is supported.");
+                std::process::exit(1);
+            }
+        }
+        #[cfg(not(feature = "remote"))]
+        {
+            eprintln!("Error: this binary was not compiled with the 'remote' feature enabled. Hence, remote operations are not supported.");
             std::process::exit(1);
         }
-    }
-
-    // Handle `verify-attribution` command
-    if let Some(matches) = matches.subcommand_matches("verify-attribution") {
-        let model_passport_path =
-            Path::new(matches.get_one::<String>("model_passport_path").unwrap());
-        let attribution_certificate_path = Path::new(
-            matches
-                .get_one::<String>("attribution_certificate_path")
-                .unwrap(),
-        );
-
-        // Call the verify function
-        local::verify_attribution(model_passport_path, attribution_certificate_path)
-            .await
-            .map_err(|err| format!("Error verifying attribution: {err}"))?;
+    } else {
+        eprintln!("Error: No valid subcommand provided. Use '--help' to see available commands.");
+        std::process::exit(1);
     }
 
     Ok(())
