+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "local")]
+#[cfg(feature = "local")]
 pub mod local;
 #[cfg(feature = "remote")]
 pub mod remote;
